Refactor tests involving login
