+++ src/test/java/com/google/sps/EventServletTest.java
@@ -26,93 +26,43 @@ import com.google.appengine.api.datastore.Query;
 import com.google.appengine.api.users.UserServiceFactory;
 import com.google.appengine.tools.development.testing.LocalDatastoreServiceTestConfig;
 import com.google.appengine.tools.development.testing.LocalServiceTestHelper;
-import com.google.gson.Gson;
-import com.google.sps.servlets.AuthServlet;
 import com.google.sps.servlets.EventServlet;
 import java.io.IOException;
-import java.io.PrintWriter;
-import java.io.StringWriter;
-import java.net.MalformedURLException;
-import java.net.URL;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.powermock.api.mockito.PowerMockito;
 import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.junit4.PowerMockRunner;
 
 /** */
 @RunWith(PowerMockRunner.class)
-@PrepareForTest({URL.class, UserServiceFactory.class})
+@PrepareForTest(UserServiceFactory.class)
 public final class EventServletTest {
   private final LocalServiceTestHelper helper =
       new LocalServiceTestHelper(new LocalDatastoreServiceTestConfig());
-  private final Gson gson = new Gson();
   private EventServlet testEventServlet;
-  private AuthServlet testAuthServlet;
-  private MockedUserService mockService;
-
-  private String activeUrl;
-
-  /**
-   * Use the current url to login/logout
-   *
-   * @param email If logging in, will log into this user's account.
-   */
-  private void toggleLogin(String email) throws MalformedURLException, IOException {
-    URL mockurl = PowerMockito.mock(URL.class);
-    when(mockurl.openConnection())
-        .thenReturn(mockService.evaluateURL(AuthServletTest.makeLoginURL(activeUrl, email)));
-    mockurl.openConnection();
-
-    HttpServletRequest request = mock(HttpServletRequest.class);
-    HttpServletResponse response = mock(HttpServletResponse.class);
-    StringWriter out = new StringWriter();
-    PrintWriter writer = new PrintWriter(out);
-    when(response.getWriter()).thenReturn(writer);
-
-    testAuthServlet.doGet(request, response);
-    out.flush();
-    LoginObject result = gson.fromJson(out.toString(), LoginObject.class);
-    activeUrl = result.url;
-  }
 
   @Before
   public void setUp() throws IOException {
     helper.setUp();
-    PowerMockito.mockStatic(UserServiceFactory.class);
-    mockService = new MockedUserService();
-    when(UserServiceFactory.getUserService()).thenReturn(mockService);
     testEventServlet = new EventServlet();
-    testAuthServlet = new AuthServlet();
-
-    // get the initial login url
-    HttpServletRequest request = mock(HttpServletRequest.class);
-    HttpServletResponse response = mock(HttpServletResponse.class);
-    StringWriter out = new StringWriter();
-    PrintWriter writer = new PrintWriter(out);
-    when(response.getWriter()).thenReturn(writer);
-    testAuthServlet.doGet(request, response);
-    out.flush();
-
-    LoginObject result = gson.fromJson(out.toString(), LoginObject.class);
-    activeUrl = result.url;
+    TestingUtil.setUp();
   }
 
   @After
   public void tearDown() {
     helper.tearDown();
-    activeUrl = "";
+    TestingUtil.tearDown();
   }
 
   @Test
   public void postOneEventToDatastore() throws IOException {
     HttpServletRequest request = mock(HttpServletRequest.class);
     HttpServletResponse response = mock(HttpServletResponse.class);
-    toggleLogin("test@example.com");
+    TestingUtil.toggleLogin("test@example.com");
 
     // Add a mock request to pass as a parameter to doPost.
     when(request.getParameter("event-name")).thenReturn("Lake Clean Up");
@@ -138,7 +88,7 @@ public final class EventServletTest {
   public void postMultipleEventsToDatastore() throws IOException {
     HttpServletRequest request = mock(HttpServletRequest.class);
     HttpServletResponse response = mock(HttpServletResponse.class);
-    toggleLogin("test@example.com");
+    TestingUtil.toggleLogin("test@example.com");
 
     // Add a mock request to pass as a parameter to doPost.
     when(request.getParameter("event-name")).thenReturn("Lake Clean Up");
@@ -167,7 +117,7 @@ public final class EventServletTest {
     HttpServletRequest request = mock(HttpServletRequest.class);
     HttpServletResponse response = mock(HttpServletResponse.class);
     String creatorEmail = "test@example.com";
-    toggleLogin(creatorEmail);
+    TestingUtil.toggleLogin(creatorEmail);
 
     // Add a mock request to pass as a parameter to doPost.
     when(request.getParameter("event-name")).thenReturn("Lake Clean Up");
@@ -212,7 +162,7 @@ public final class EventServletTest {
   public void postEventWithEmptyOptionalFields() throws IOException {
     HttpServletRequest request = mock(HttpServletRequest.class);
     HttpServletResponse response = mock(HttpServletResponse.class);
-    toggleLogin("test@example.com");
+    TestingUtil.toggleLogin("test@example.com");
 
     // This mock request does not include optional fields end-time and cover-photo.
     when(request.getParameter("event-name")).thenReturn("Lake Clean Up");
@@ -261,11 +211,5 @@ public final class EventServletTest {
       DatastoreService ds = DatastoreServiceFactory.getDatastoreService();
       assertEquals(0, ds.prepare(new Query("Event")).countEntities());
     }
-  }
-
-  /* the LoginObject structure used by AuthServlet */
-  private static class LoginObject {
-    private boolean loggedIn;
-    private String url;
   }
 }
