+++ src/test/java/com/google/sps/InteractionsTest.java
@@ -19,14 +19,9 @@ import static org.mockito.Mockito.*;
 import com.google.appengine.api.users.UserServiceFactory;
 import com.google.appengine.tools.development.testing.LocalDatastoreServiceTestConfig;
 import com.google.appengine.tools.development.testing.LocalServiceTestHelper;
-import com.google.gson.Gson;
-import com.google.sps.servlets.AuthServlet;
 import com.google.sps.servlets.SurveyServlet;
 import java.io.IOException;
-import java.io.PrintWriter;
-import java.io.StringWriter;
 import java.net.MalformedURLException;
-import java.net.URL;
 import java.util.Map;
 import java.util.TreeMap;
 import javax.servlet.http.HttpServletRequest;
@@ -35,48 +30,19 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.powermock.api.mockito.PowerMockito;
 import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.junit4.PowerMockRunner;
 
 /** */
 @RunWith(PowerMockRunner.class)
-@PrepareForTest({URL.class, UserServiceFactory.class})
+@PrepareForTest(UserServiceFactory.class)
 public final class InteractionsTest {
   private final LocalServiceTestHelper helper =
       new LocalServiceTestHelper(new LocalDatastoreServiceTestConfig());
-  private final Gson gson = new Gson();
-  private AuthServlet testAuthServlet;
   private SurveyServlet testSurveyServlet;
-  private MockedUserService mockService;
-
-  private String activeUrl;
-
-  /**
-   * Use the current url to login/logout
-   *
-   * @param email If logging in, will log into this user's account.
-   */
-  private void toggleLogin(String email) throws MalformedURLException, IOException {
-    URL mockurl = PowerMockito.mock(URL.class);
-    when(mockurl.openConnection())
-        .thenReturn(mockService.evaluateURL(AuthServletTest.makeLoginURL(activeUrl, email)));
-    mockurl.openConnection();
-
-    HttpServletRequest request = mock(HttpServletRequest.class);
-    HttpServletResponse response = mock(HttpServletResponse.class);
-    StringWriter out = new StringWriter();
-    PrintWriter writer = new PrintWriter(out);
-    when(response.getWriter()).thenReturn(writer);
-
-    testAuthServlet.doGet(request, response);
-    out.flush();
-    LoginObject result = gson.fromJson(out.toString(), LoginObject.class);
-    activeUrl = result.url;
-  }
 
   private void takeSurvey(String email) throws MalformedURLException, IOException {
-    toggleLogin(email);
+    TestingUtil.toggleLogin(email);
 
     HttpServletRequest request = mock(HttpServletRequest.class);
     HttpServletResponse response = mock(HttpServletResponse.class);
@@ -87,35 +53,20 @@ public final class InteractionsTest {
     when(request.getParameter("LGBTQ+")).thenReturn("4");
 
     testSurveyServlet.doPost(request, response);
-    toggleLogin(email);
+    TestingUtil.toggleLogin(email);
   }
 
   @Before
   public void setUp() throws IOException {
     helper.setUp();
-    PowerMockito.mockStatic(UserServiceFactory.class);
-    mockService = new MockedUserService();
-    when(UserServiceFactory.getUserService()).thenReturn(mockService);
-    testAuthServlet = new AuthServlet();
     testSurveyServlet = new SurveyServlet();
-
-    // get the initial login url
-    HttpServletRequest request = mock(HttpServletRequest.class);
-    HttpServletResponse response = mock(HttpServletResponse.class);
-    StringWriter out = new StringWriter();
-    PrintWriter writer = new PrintWriter(out);
-    when(response.getWriter()).thenReturn(writer);
-    testAuthServlet.doGet(request, response);
-    out.flush();
-
-    LoginObject result = gson.fromJson(out.toString(), LoginObject.class);
-    activeUrl = result.url;
+    TestingUtil.setUp();
   }
 
   @After
   public void tearDown() {
     helper.tearDown();
-    activeUrl = "";
+    TestingUtil.tearDown();
   }
 
   @Test
@@ -139,7 +90,7 @@ public final class InteractionsTest {
     // call Interactions.getInterestMetrics() on nonexistent user
 
     String email = "test@example.com";
-    toggleLogin(email);
+    TestingUtil.toggleLogin(email);
 
     HttpServletRequest request = mock(HttpServletRequest.class);
     HttpServletResponse response = mock(HttpServletResponse.class);
@@ -151,11 +102,5 @@ public final class InteractionsTest {
     testSurveyServlet.doPost(request, response);
 
     assertEquals(null, Interactions.getInterestMetrics("other@example.com"));
-  }
-
-  /* the LoginObject structure used by AuthServlet */
-  private static class LoginObject {
-    private boolean loggedIn;
-    private String url;
   }
 }
